Выносит фильтрацию в отдельный модуль
