+++ js/gallery.js
@@ -1,21 +1,9 @@
 import {openBigPicture} from './big-picture.js';
-import { getData } from './server.js';
-import { createUniqueIds, debounce } from './utils.js';
-
-const RANDOM_PHOTO_COUNT = 10;
-
-const GallerySortingAction = {
-  'filter-default': sortByDefault,
-  'filter-random': sortByRandom,
-  'filter-discussed': sortByDiscussed
-};
 
 const templateThumbnail = document.querySelector('#picture').content.querySelector('.picture');
 const picturesContainer = document.querySelector('.pictures');
 const thumbnailsContainer = document.querySelector('.pictures');
-const filtersSection = document.querySelector('.img-filters');
-const filtersForm = filtersSection.querySelector('.img-filters__form');
-let activeFilter = 'filter-default';
+
 let userPhotos = [];
 
 const createThumbnailItem = ({url, description, likes, comments, id}) => {
@@ -31,7 +19,7 @@ const createThumbnailItem = ({url, description, likes, comments, id}) => {
   return thumbnail;
 };
 
-const renderGallery = (photos) => {
+export const renderGallery = (photos) => {
   const fragment = document.createDocumentFragment();
 
   photos.forEach((photo) => {
@@ -42,46 +30,13 @@ const renderGallery = (photos) => {
   picturesContainer.append(fragment);
 };
 
-const initFilters = () => filtersSection.classList.remove('img-filters--inactive');
-
-getData()
-  .then((photos) => {
-    userPhotos = photos;
-    renderGallery(photos);
-  })
-  .then(() => {
-    initFilters();
-  });
-
-const debouncedFilterGallery = debounce((newFilterId) => {
-  if (newFilterId === activeFilter) {
-    return;
-  }
-
-  activeFilter = newFilterId;
-
-  const action = GallerySortingAction[newFilterId];
-  action();
-});
-
-const onGalleryFiltersClick = ({target}) => {
-  if (!target.classList.contains('img-filters__button')) {
-    return;
-  }
-
-  const activeButton = filtersForm.querySelector('.img-filters__button--active');
-
-  if (activeButton === target) {
-    return;
-  }
-
-  activeButton?.classList.remove('img-filters__button--active');
-  target.classList.add('img-filters__button--active');
-
-  debouncedFilterGallery(target.id);
+export const initGallery = (photos) => {
+  userPhotos = photos;
+  renderGallery(photos);
 };
 
-filtersForm.addEventListener('click', onGalleryFiltersClick);
+export const getUserPhotos = () => userPhotos;
+
 
 thumbnailsContainer.addEventListener('click', (evt) => {
   const thumbnail = evt.target.closest('.picture[data-id]');
@@ -99,26 +54,3 @@ thumbnailsContainer.addEventListener('click', (evt) => {
   openBigPicture(photoData);
 });
 
-const clearPhotos = () => picturesContainer.querySelectorAll('.picture').forEach((thumbnail) => thumbnail.remove());
-
-function sortByDiscussed () {
-  clearPhotos();
-
-  const sortedPhotos = userPhotos.toSorted((photoA, photoB) => photoB.comments.length - photoA.comments.length);
-
-  renderGallery(sortedPhotos);
-}
-
-function sortByDefault () {
-  clearPhotos();
-  renderGallery(userPhotos);
-}
-
-function sortByRandom () {
-  clearPhotos();
-
-  const uniqueIds = createUniqueIds(userPhotos[0].id, userPhotos[userPhotos.length - 1].id, RANDOM_PHOTO_COUNT);
-  const sortedPhotos = userPhotos.filter((item) => uniqueIds.has(item.id));
-
-  renderGallery(sortedPhotos);
-}
